Add `isAppActive` on `BackgroundTaskScheduler`
