+++ Sources/StreamChat/WebSocketClient/WebSocketClient_Tests.swift
@@ -528,39 +528,6 @@ extension WebSocketEngineError: Equatable {
     }
 }
 
-class MockBackgroundTaskScheduler: BackgroundTaskScheduler {
-    var beginBackgroundTask_called: Bool = false
-    var beginBackgroundTask_expirationHandler: (() -> Void)?
-    var beginBackgroundTask_returns: Bool = true
-    func beginTask(expirationHandler: (() -> Void)?) -> Bool {
-        beginBackgroundTask_called = true
-        beginBackgroundTask_expirationHandler = expirationHandler
-        return beginBackgroundTask_returns
-    }
-
-    var endBackgroundTask_called: Bool = false
-    func endTask() {
-        endBackgroundTask_called = true
-    }
-
-    var startListeningForAppStateUpdates_called: Bool = false
-    var startListeningForAppStateUpdates_onBackground: (() -> Void)?
-    var startListeningForAppStateUpdates_onForeground: (() -> Void)?
-    func startListeningForAppStateUpdates(
-        onEnteringBackground: @escaping () -> Void,
-        onEnteringForeground: @escaping () -> Void
-    ) {
-        startListeningForAppStateUpdates_called = true
-        startListeningForAppStateUpdates_onBackground = onEnteringBackground
-        startListeningForAppStateUpdates_onForeground = onEnteringForeground
-    }
-    
-    var stopListeningForAppStateUpdates_called: Bool = false
-    func stopListeningForAppStateUpdates() {
-        stopListeningForAppStateUpdates_called = true
-    }
-}
-
 class WebSocketPingControllerMock: WebSocketPingController {
     var connectionStateDidChange_connectionStates: [WebSocketConnectionState] = []
     var pongRecievedCount = 0
